requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
exceptiongroup==1.2.2
factory_boy==3.3.3
Faker==37.1.0
fastapi==0.115.12
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==25.4.2
geventhttpclient==2.3.3
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.34.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
packaging==25.0
passlib==1.7.4
pastel==0.2.1
pluggy==1.5.0
poethepoet==0.34.0
psutil==7.0.0
psycopg2==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
python-decouple==3.8
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
rich==14.0.0
ruff==0.11.8
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tomli==2.2.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
Werkzeug==3.1.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
